+++ synapse-core/src/main/java/de/otto/synapse/endpoint/sender/MessageSenderEndpoint.java
@@ -1,12 +1,6 @@
 package de.otto.synapse.endpoint.sender;
 
 import de.otto.synapse.endpoint.MessageEndpoint;
-import de.otto.synapse.message.Message;
-import de.otto.synapse.translator.MessageFormat;
-
-import javax.annotation.Nonnull;
-import java.util.concurrent.CompletableFuture;
-import java.util.stream.Stream;
 
 /**
  * Endpoint that is used by an application to send messages to a messaging channel.
@@ -18,29 +12,6 @@ import java.util.stream.Stream;
  * @see <a href="http://www.enterpriseintegrationpatterns.com/patterns/messaging/MessageEndpoint.html">EIP: Message Endpoint</a>
  */
 
-public interface MessageSenderEndpoint extends MessageEndpoint {
+public interface MessageSenderEndpoint extends MessageSender, MessageEndpoint {
 
-    /**
-     * Send a single {@link Message} to the channel.
-     *
-     * @param message the message
-     * @param <T> the type of the message payload
-     * @return a CompletableFuture that can be used for logging, error handling, waiting for completion, etc.
-     */
-    <T> CompletableFuture<Void> send(@Nonnull Message<T> message);
-
-    /**
-     * Send a batch of {@link Message messages} to the channel.
-     *
-     * @param batch the batch of messages
-     * @param <T> the type of the message payload
-     * @return a CompletableFuture that can be used for logging, error handling, waiting for completion, etc.
-     */
-    <T> CompletableFuture<Void> sendBatch(@Nonnull Stream<Message<T>> batch);
-
-    /**
-     * Returns the message format of the sender
-     * @return messageFormat the message format
-     */
-    MessageFormat getMessageFormat();
 }