+++ synapse-examples/example-messagequeue-consumer/src/main/java/de/otto/synapse/example/consumer/SampleProducer.java
@@ -1,6 +1,6 @@
 package de.otto.synapse.example.consumer;
 
-import de.otto.synapse.endpoint.sender.MessageSenderEndpoint;
+import de.otto.synapse.endpoint.sender.MessageSender;
 import de.otto.synapse.example.consumer.configuration.MyServiceProperties;
 import de.otto.synapse.example.consumer.payload.BananaPayload;
 import de.otto.synapse.example.consumer.payload.ProductPayload;
@@ -23,9 +23,9 @@ import static java.util.Arrays.asList;
 public class SampleProducer {
 
     @Autowired
-    private MessageSenderEndpoint bananaMessageSender;
+    private MessageSender bananaMessageSender;
     @Autowired
-    private MessageSenderEndpoint productMessageSender;
+    private MessageSender productMessageSender;
 
     @Scheduled(initialDelay = 1000L, fixedRate = 10000L)
     protected void produceBananaSampleData() {